[#17] index.py works, search.py works.
