+++ requirements.txt
@@ -6,6 +6,8 @@ click==8.0.3
 configparser==5.0.2
 docker-pycreds==0.4.0
 elasticsearch==7.15.0
+elasticsearch-dsl==7.4.0
+gcloud==0.18.3
 gitdb==4.0.7
 GitPython==3.1.24
 google-api-core==2.1.1
@@ -18,7 +20,8 @@ googleapis-common-protos==1.53.0
 httplib2==0.20.1
 idna==3.3
 importlib-metadata==4.8.1
-joblib==1.1.0
+joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
+more-itertools==8.10.0
 numpy==1.21.2
 oauth2client==4.1.3
 pandas==1.3.3
@@ -35,8 +38,8 @@ pytz==2021.3
 PyYAML==6.0
 requests==2.26.0
 rsa==4.7.2
-scikit-learn==1.0
-scipy==1.7.1
+scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1632611447519/work
+scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1633644994244/work
 sentry-sdk==1.4.3
 shortuuid==1.0.1
 six==1.16.0
@@ -44,7 +47,7 @@ smmap==4.0.0
 soupsieve==2.2.1
 subprocess32==3.5.4
 termcolor==1.1.0
-threadpoolctl==3.0.0
+threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1633102299089/work
 tqdm==4.62.3
 typing-extensions==3.10.0.2
 urllib3==1.26.7
