+++ storyteller/explore/explore_doc.py
@@ -1,0 +1,10 @@
+
+from storyteller.docs import GKDoc
+
+
+def main():
+    print(GKDoc)
+
+
+if __name__ == '__main__':
+    main()
