+++ storyteller/main/authorize.py
@@ -1,0 +1,24 @@
+"""
+Authorize the user to use google cloud & wandb, once and for all,
+just by running this python script.
+Pre-requisite: fill up the .env file with your gcloud & wandb credentials!
+"""
+
+
+# --- the helper --- #
+class Authorizer:
+    """
+    Supports storyteller/main/authorize.py
+    """
+
+    def __call__(self, *args, **kwargs):
+        pass
+
+
+# --- the script --- #
+def main():
+    pass
+
+
+if __name__ == '__main__':
+    main()
